feat(unity): 밸런스 개편 미러링 — 웨이브 모델·프로브·보스·XP
web 커밋 625987b의 C# 미러:
- Balance.cs: ExpectedBoardDps/TargetClearSeconds/EnemyHP(모델 역산),
  몹 수 20~36, 스폰 0.18, 프로브 100/1.5, 보스 HP ×2.5,
  보스 보상 [5,10,18,32,55,90] [프로토] (원본 표는 주석 보존), waveReward 10+3r
- HeroData.cs: XP_PER_MOB 0.65f (float), BOSS_HARMLESS_MAX_LEVEL=3 — Lv3까지
  접촉 무해, BossDamage가 0을 반환
- Hero.Xp / GainXp / GrantXp int → float (XP 소수화), HUD 표시는 내림
- 스텁에 Mathf.FloorToInt 보강. Core/View 타입체크 0에러

+++ unity/Assets/Scripts/Core/Game.cs
@@ -447,7 +447,7 @@ namespace GodTD.Core
         }
 
         /// <summary>경험치는 타워가 잡든 영웅이 잡든 들어온다. 레벨업 시 증강 선택을 띄운다.</summary>
-        void GrantXp(int amount)
+        void GrantXp(float amount)
         {
             var hero = Hero;
             int levels = hero.GainXp(amount);
